Converted frontend from js to ts
