+++ frontend/src/pages/Admin.tsx
@@ -1,0 +1,11 @@
+import React from 'react'
+
+const Admin = () => {
+  return (
+    <div>
+      This is admin
+    </div>
+  )
+}
+
+export default Admin
