--- frontend/src/pages/Home.jsx
@@ -1,9 +1,0 @@
-import React from 'react'
-
-const Home = () => {
-    return (
-        <div>Home Page</div>
-    )
-}
-
-export default Home
